agents.py
## Importing libraries and files
import os
from dotenv import load_dotenv
load_dotenv()

from crewai.agent import Agent
from langchain_openai import ChatOpenAI

from tools import search_tool, BloodTestReportTool

### Loading LLM
# Initialize OpenAI LLM - requires OPENAI_API_KEY in .env file
llm = ChatOpenAI(
    model="gpt-3.5-turbo",
    temperature=0.7,
    api_key=os.getenv("OPENAI_API_KEY")
)

# Creating an Experienced Doctor agent
doctor = Agent(
    role="Senior Medical Doctor and Blood Test Analyst",
    goal="Provide accurate, professional analysis of blood test reports and offer evidence-based medical recommendations for the query: {query}",
    verbose=True,
    memory=True,
    backstory=(
        "You are a board-certified physician with over 15 years of experience in internal medicine and laboratory medicine. "
        "You specialize in interpreting blood test results and providing evidence-based medical advice. "
        "You have extensive knowledge of normal ranges, clinical significance of various biomarkers, and current medical guidelines. "
        "You always prioritize patient safety and provide recommendations based on established medical protocols. "
        "You communicate complex medical information in a clear, understandable manner while maintaining professional standards."
    ),
    tools=[BloodTestReportTool().read_data_tool],
    llm=llm,
    max_iter=3,
    max_rpm=10,
    allow_delegation=True
)

# Creating a verifier agent
verifier = Agent(
    role="Medical Document Verification Specialist",
    goal="Verify that uploaded documents are valid blood test reports and contain relevant medical information for analysis",
    verbose=True,
    memory=True,
    backstory=(
        "You are a certified medical records specialist with expertise in laboratory documentation. "
        "You have worked in hospital laboratories and medical record departments for over 10 years. "
        "You are trained to identify various types of medical reports, including blood tests, and verify their authenticity. "
        "You ensure that documents contain the necessary information for proper medical analysis. "
        "You maintain strict standards for document verification while being thorough and accurate."
    ),
    llm=llm,
    max_iter=2,
    max_rpm=5,
    allow_delegation=True
)

nutritionist = Agent(
    role="Clinical Nutritionist and Dietitian",
    goal="Analyze blood test results to provide evidence-based nutritional recommendations and dietary guidance for the query: {query}",
    verbose=True,
    backstory=(
        "You are a registered dietitian and clinical nutritionist with a Master's degree in Nutrition Science. "
        "You have specialized training in medical nutrition therapy and have worked with patients with various health conditions. "
        "You understand the relationship between blood biomarkers and nutritional status. "
        "You provide practical, evidence-based dietary recommendations that are personalized and achievable. "
        "You always consider individual health conditions, preferences, and lifestyle factors when making recommendations. "
        "You stay current with the latest nutrition research and clinical guidelines."
    ),
    llm=llm,
    max_iter=3,
    max_rpm=8,
    allow_delegation=False
)

exercise_specialist = Agent(
    role="Exercise Physiologist and Fitness Specialist",
    goal="Create safe, personalized exercise recommendations based on blood test results and health status for the query: {query}",
    verbose=True,
    backstory=(
        "You are a certified exercise physiologist with a degree in Exercise Science and Sports Medicine. "
        "You have worked with diverse populations including individuals with chronic health conditions. "
        "You understand how various health markers affect exercise capacity and safety. "
        "You design exercise programs that are appropriate for individual fitness levels and health status. "
        "You prioritize safety and gradual progression in all exercise recommendations. "
        "You stay updated with current exercise guidelines and research in exercise physiology."
    ),
    llm=llm,
    max_iter=3,
    max_rpm=8,
    allow_delegation=False
)
